% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_uncertainty_init.R
\name{boot_prob_marg}
\alias{boot_prob_marg}
\title{Univariate probabilities estimated through bootstrap}
\usage{
boot_prob_marg(name_marg, ci_par, quantile, B_m, ci_level, thresh_GPD = FALSE)
}
\arguments{
\item{name_marg}{character string specifying the name of the distribution family.}

\item{ci_par}{list of the estimated parameters and their confidence interval}

\item{quantile}{quantile value.}

\item{B_m}{number of samples}

\item{ci_level}{the confidence level required.}

\item{thresh_GPD}{threshold value used for the GPD distribution.}
}
\value{
vector of dimension B_m containing probabilities estimated through bootstrap sampling.
}
\description{
Generate a bootstrap sample of probabilities
}
