% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe_contrib_from_prob.R
\name{contribution}
\alias{contribution}
\title{Contribution metric}
\usage{
contribution(proba, p0, prob_margdep)
}
\arguments{
\item{proba}{CE probability when one or two statistical components evolve.}

\item{p0}{CE probability during the reference period.}

\item{prob_margdep}{CE probability when the three statistical components evolve.}
}
\value{
the contribution value.
}
\description{
Give the percentage of each statistical component (marginal1, marginal2, dependence) change to the overall signal change.
}
