% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe_contrib_from_prob.R
\name{detect_poe}
\alias{detect_poe}
\title{Periods of emergence (PoE) detection}
\usage{
detect_poe(
  proba,
  low_bound,
  up_bound,
  y_start,
  y_end,
  length_sliding_window,
  step_sliding_window
)
}
\arguments{
\item{proba}{CE probability when one, two or the three statistical components evolve.}

\item{low_bound}{lower bound of the natural variability (confidence interval of the probability estimated during the reference period).}

\item{up_bound}{upper bound of the natural variability (confidence interval of the probability estimated during the reference period).}

\item{y_start}{first year of the whole studied period.}

\item{y_end}{last year of the whole studied period.}

\item{length_sliding_window}{number of years of the sliding window.}

\item{step_sliding_window}{number of years separating two sliding windows.}
}
\value{
a list containing PoE features (number of PoEs, their duration, their starting year, the potential ToE).
}
\description{
Detect PoEs, periods during which the signal exceeds the natural variability, either below the lower bound or above the upper bound.
}
