% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_uncertainty_init.R
\name{determine_ci_par_marg_boot}
\alias{determine_ci_par_marg_boot}
\title{Uncertainty of the marginal parameters ("bootstrap")}
\usage{
determine_ci_par_marg_boot(
  name_marg,
  par,
  n_sample,
  nb_pt,
  ci_level,
  thresh_GPD = NaN
)
}
\arguments{
\item{name_marg}{character string specifying the name of the distribution family.}

\item{par}{list of parameters.}

\item{n_sample}{number of samples for the bootstrap.}

\item{nb_pt}{number of points to simulate.}

\item{ci_level}{the confidence level required.}

\item{thresh_GPD}{threshold value to be used for GPD distribution. Defaults to NaN.}
}
\value{
vector of 3 values : initial parameters (input) and the confidence interval associated to this value estimated with a bootstrap approach.
}
\description{
Give the confidence interval of the marginal parameters with a bootstrap approach.
}
