% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_uncertainty_init.R
\name{determine_ci_par_marg_confint}
\alias{determine_ci_par_marg_confint}
\title{Uncertainty of the marginal parameters ("ml")}
\usage{
determine_ci_par_marg_confint(name_marg, data_x, ci_level, thresh_GPD = NaN)
}
\arguments{
\item{name_marg}{character string specifying the name of the distribution family.}

\item{data_x}{numeric vector.}

\item{ci_level}{the confidence level required.}

\item{thresh_GPD}{threshold value to be used for GPD distribution. Defaults to NaN.}
}
\value{
vector of 3 values : initial parameters (input) and the confidence interval associated to this value estimated with the confint() function from the package 'stats'.
}
\description{
Give the confidence interval of the marginal parameters.
}
