% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_uncertainty_init.R
\name{determine_ci_theta_boot}
\alias{determine_ci_theta_boot}
\title{Uncertainty of the copula parameter ("bootstrap")}
\usage{
determine_ci_theta_boot(name_cop, theta, n_sample, nb_pt, ci_level)
}
\arguments{
\item{name_cop}{character string specifying the copula name.}

\item{theta}{copula parameter.}

\item{n_sample}{number of samples for the bootstrap.}

\item{nb_pt}{number of points to simulate.}

\item{ci_level}{the confidence level required.}
}
\value{
vector of 3 values : theta (input) and the confidence interval associated to this value estimated with a bootstrap approach.
}
\description{
Give the confidence interval of the copula parameter with a bootstrap approach.
}
