% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_uncertainty_init.R
\name{determine_ci_theta_confint}
\alias{determine_ci_theta_confint}
\title{Uncertainty of the copula parameter ("ml")}
\usage{
determine_ci_theta_confint(cop_object, data_U, ci_level = ci_level)
}
\arguments{
\item{cop_object}{copula object.}

\item{data_U}{numeric vector of pseudo-observations.}

\item{ci_level}{the confidence level required.}
}
\value{
vector of 3 values : theta (input) and the confidence interval associated to this value estimated with the confint() function from the package 'stats'.
}
\description{
Give the confidence interval of the copula parameter.
}
