% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{determine_par_marg}
\alias{determine_par_marg}
\title{Estimating marginal parameters}
\usage{
determine_par_marg(data_x, name_family, thresh_GPD = NaN)
}
\arguments{
\item{data_x}{numeric vector.}

\item{name_family}{character string specifying the distribution family name.}

\item{thresh_GPD}{threshold value to be used for GPD distribution. Defaults to NaN.}
}
\value{
a list with the estimated parameters (par) and the loglikelihood (logLik)
}
\description{
Parameter estimation of univariate distributions, i.e., fitting of a distribution to observations.
}
