% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{determine_theta_copula}
\alias{determine_theta_copula}
\title{Estimating copula parameter}
\usage{
determine_theta_copula(data_U, name_family)
}
\arguments{
\item{data_U}{numeric vector of pseudo-observations.}

\item{name_family}{character string specifying the copula name.}
}
\value{
a list with the estimated parameters (par) and the loglikelihood (logLik)
}
\description{
Parameter estimation of copulas, i.e., fitting of a copula model to bivariate “pseudo” observations.
}
