% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_uncertainty_init.R
\name{determine_uncertainty_init}
\alias{determine_uncertainty_init}
\title{Uncertainty of the joint exceedance probability during the reference period}
\usage{
determine_uncertainty_init(
  series1,
  series2,
  y_start,
  y_end,
  length_sliding_window,
  threshold1,
  threshold2,
  res_distr_max_minAIC,
  ci_level = 0.68,
  method = "ml",
  B = 500
)
}
\arguments{
\item{series1}{numeric vector.}

\item{series2}{numeric vector.}

\item{y_start}{first year of the whole studied period.}

\item{y_end}{last year of the whole studied period.}

\item{length_sliding_window}{number of years of the sliding window.}

\item{threshold1}{threshold for series1, used to compute the joint exceedance probability (the probability that variable1 exceeds threshold1 and variable2 exceeds threshold2).}

\item{threshold2}{threshold for series2, used to compute the joint exceedance probability (the probability that variable1 exceeds threshold1 and variable2 exceeds threshold2).}

\item{res_distr_max_minAIC}{output of the function sel_distr_max_minAIC(), giving the fit features of each component (marginal 1, marginal 2 and copula).}

\item{ci_level}{the confidence level required. Defaults to 0.68.}

\item{method}{character string, either "ml" (default) or "bootstrap.}

\item{B}{number of iterations for the bootstrap. Defaults to 500.}
}
\value{
vector of 3 values : the estimated CE probability during the reference period and the confidence interval associated to this value.
}
\description{
Compute uncertainty of CE probability with two methods ("ml" and "bootstrap").
}
