% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{fit_distr}
\alias{fit_distr}
\title{Fitting marginals and copula to data}
\usage{
fit_distr(
  series1,
  series2,
  y_start,
  y_end,
  length_sliding_window,
  step_sliding_window,
  list_marg1 = c("normal", "GEV", "gamma", "log-normal", "GPD"),
  list_marg2 = c("normal", "GEV", "gamma", "log-normal", "GPD"),
  list_cop = c("Normal", "Clayton", "Gumbel", "Joe", "Frank"),
  thresh1_GPD = NaN,
  thresh2_GPD = NaN
)
}
\arguments{
\item{series1}{numeric vector.}

\item{series2}{numeric vector.}

\item{y_start}{first year of the whole studied period.}

\item{y_end}{last year of the whole studied period.}

\item{length_sliding_window}{number of years of the sliding window.}

\item{step_sliding_window}{number of years separating two sliding windows.}

\item{list_marg1}{character vector specifying the names of distribution family to be tested for series1.
Defaults to c("normal","GEV","gamma","log-normal",GPD").}

\item{list_marg2}{character vector specifying the names of distribution family to be tested for series2.
Defaults to c("normal","GEV","gamma","log-normal",GPD").}

\item{list_cop}{character vector specifying the names of copula family to be tested.
Defaults to c("Normal","Clayton","Gumbel","Joe","Frank").}

\item{thresh1_GPD}{threshold value to be used if GPD family is tested for series 1 (i.e. included in list_marg1).
Defaults to NaN.}

\item{thresh2_GPD}{threshold value to be used if GPD family is tested for series 2 (i.e. included in list_marg2).
Defaults to NaN.}
}
\value{
a list with the estimated parameters, the loglikelihood value, AIC and BIC for each tested distribution family of each component (marginal 1, marginal 2, copula).
}
\description{
Give the fit features for several distributions in order to be able to choose the best marginal and copula family depending on the criteria (loglikelihood, AIC, BIC).
}
