% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{middle_sliding_window}
\alias{middle_sliding_window}
\title{Middle of each sliding window}
\usage{
middle_sliding_window(label)
}
\arguments{
\item{label}{vector of character string specifying each sliding period.}
}
\value{
numeric vector specifying the midpoint year of each period.
}
\description{
Give the middle year of each period.
}
\examples{
label = c("1950_1969","1951_1970","1952_1971","1953_1972")
years = middle_sliding_window(label)

}
