% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_marg}
\alias{p_marg}
\alias{q_marg}
\title{Density distribution and quantile function}
\usage{
p_marg(name_marg, par, quantile, thresh_GPD = NaN)
}
\arguments{
\item{name_marg}{character string specifying the distribution family name. Examples : "normal","log-normal","GEV","GPD","gamma".}

\item{par}{list of parameters.}

\item{quantile}{quantile value.}

\item{thresh_GPD}{threshold value for GPD distribution. Default to NaN.}
}
\value{
p_marg gives the probability associated to the quantile, q_marg gives the quantile associated to the probability
}
\description{
Density distribution and quantile function for the following cases : "normal","log"normal","GEV","GPD","gamma".
}
