% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe_contrib_from_prob.R
\name{poe_contrib_from_prob}
\alias{poe_contrib_from_prob}
\title{Detect emergence (poe and toe) and quantify the contributions of each statistical component (marginals and dependence)}
\usage{
poe_contrib_from_prob(
  res_5prob,
  res_ci,
  y_start,
  y_end,
  length_sliding_window,
  step_sliding_window
)
}
\arguments{
\item{res_5prob}{output of the function compute_5prob() giving CE probabilities when one, two or the three statistical components (marginal 1, marginal 2 and the dependence) vary while the others are held constant.}

\item{res_ci}{output of the function determine_uncertainty_init() giving the probability associated with the natural variability.}

\item{y_start}{first year of the whole studied period.}

\item{y_end}{last year of the whole studied period.}

\item{length_sliding_window}{number of years of the sliding window.}

\item{step_sliding_window}{number of years separating two sliding windows.}
}
\value{
a list containing the features associated with periods of emergence (poe) and the values of each statistical's component contribution.
}
\description{
Give periods of emergence (PoE) features (e.g., the number of PoEs, their duration). The signal can either emerge below the lower bound of the natural variability (PoE-low) or above the upper bound of the natural variability (PoE-up).
Give also the values of the statistical drivers associated with PoE-low and PoE-up.
}
