% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_distr_max_minAIC.R
\name{sel_distr_max_minAIC}
\alias{sel_distr_max_minAIC}
\title{Select the best family according to AIC criteria}
\usage{
sel_distr_max_minAIC(res_distr)
}
\arguments{
\item{res_distr}{outputs of the function fit_distr(), giving a list with the estimated parameters, the loglikelihood value, AIC and BIC for each tested distribution.}
}
\value{
a list with the name of the selected distribution, the estimated parameters and the loglikelihood value for each component (marginal 1, marginal 2, copula).
}
\description{
Give the distribution family that gets the highest number of times the lowest AIC values.
}
